+++ codex-rs/ext/memories/src/extension.rs
@@ -1,0 +1,84 @@
+use std::sync::Arc;
+
+use codex_core::config::Config;
+use codex_extension_api::ContextContributor;
+use codex_extension_api::ExtensionData;
+use codex_extension_api::ExtensionRegistryBuilder;
+use codex_extension_api::PromptFragment;
+use codex_extension_api::ThreadLifecycleContributor;
+use codex_extension_api::ThreadStartInput;
+use codex_extension_api::ToolContributor;
+use codex_features::Feature;
+use codex_memories_read::build_memory_tool_developer_instructions;
+use codex_utils_absolute_path::AbsolutePathBuf;
+
+use crate::local::LocalMemoriesBackend;
+use crate::tools;
+
+/// Contributes Codex memory read-path prompt context and memory read tools.
+#[derive(Clone, Copy, Debug, Default)]
+pub(crate) struct MemoriesExtension;
+
+#[derive(Clone, Debug)]
+pub(crate) struct MemoriesExtensionConfig {
+    pub(crate) enabled: bool,
+    pub(crate) codex_home: AbsolutePathBuf,
+}
+
+impl ContextContributor for MemoriesExtension {
+    fn contribute<'a>(
+        &'a self,
+        _session_store: &'a ExtensionData,
+        thread_store: &'a ExtensionData,
+    ) -> std::pin::Pin<Box<dyn std::future::Future<Output = Vec<PromptFragment>> + Send + 'a>> {
+        Box::pin(async move {
+            let Some(config) = thread_store.get::<MemoriesExtensionConfig>() else {
+                return Vec::new();
+            };
+            if !config.enabled {
+                return Vec::new();
+            }
+
+            build_memory_tool_developer_instructions(&config.codex_home)
+                .await
+                .map(PromptFragment::developer_policy)
+                .into_iter()
+                .collect()
+        })
+    }
+}
+
+impl ThreadLifecycleContributor<Config> for MemoriesExtension {
+    fn on_thread_start(&self, input: ThreadStartInput<'_, Config>) {
+        input.thread_store.insert(MemoriesExtensionConfig {
+            enabled: input.config.features.enabled(Feature::MemoryTool)
+                && input.config.memories.use_memories,
+            codex_home: input.config.codex_home.clone(),
+        });
+    }
+}
+
+impl ToolContributor for MemoriesExtension {
+    fn tools(
+        &self,
+        _session_store: &ExtensionData,
+        thread_store: &ExtensionData,
+    ) -> Vec<Arc<dyn codex_extension_api::ExtensionToolExecutor>> {
+        let Some(config) = thread_store.get::<MemoriesExtensionConfig>() else {
+            return Vec::new();
+        };
+        if !config.enabled {
+            return Vec::new();
+        }
+
+        tools::memory_tools(LocalMemoriesBackend::from_codex_home(&config.codex_home))
+    }
+}
+
+/// Installs the memories extension contributors into the extension registry.
+pub fn install(registry: &mut ExtensionRegistryBuilder<Config>) {
+    let extension = Arc::new(MemoriesExtension);
+    registry.thread_lifecycle_contributor(extension.clone());
+    registry.prompt_contributor(extension.clone());
+    registry.tool_contributor(extension);
+}
