+++ codex-rs/ext/memories/src/tools/search.rs
@@ -8,15 +8,14 @@ use schemars::JsonSchema;
 use serde::Deserialize;
 use serde_json::json;
 
-use crate::backend::DEFAULT_SEARCH_MAX_RESULTS;
-use crate::backend::MAX_SEARCH_RESULTS;
+use crate::DEFAULT_SEARCH_MAX_RESULTS;
+use crate::MAX_SEARCH_RESULTS;
+use crate::SEARCH_TOOL_NAME;
 use crate::backend::MemoriesBackend;
 use crate::backend::SearchMatchMode;
 use crate::backend::SearchMemoriesRequest;
 use crate::backend::SearchMemoriesResponse;
-use crate::local::LocalMemoriesBackend;
 
-use super::SEARCH_TOOL_NAME;
 use super::backend_error_to_function_call;
 use super::clamp_max_results;
 use super::function_tool;
@@ -39,11 +38,14 @@ struct SearchArgs {
 }
 
 #[derive(Clone)]
-pub(super) struct SearchTool {
-    pub(super) backend: LocalMemoriesBackend,
+pub(super) struct SearchTool<B> {
+    pub(super) backend: B,
 }
 
-impl ExtensionToolExecutor for SearchTool {
+impl<B> ExtensionToolExecutor for SearchTool<B>
+where
+    B: MemoriesBackend,
+{
     fn tool_name(&self) -> ToolName {
         ToolName::plain(SEARCH_TOOL_NAME)
     }
@@ -84,77 +86,5 @@ impl SearchArgs {
                 MAX_SEARCH_RESULTS,
             ),
         }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use pretty_assertions::assert_eq;
-    use serde_json::json;
-
-    use super::*;
-
-    #[test]
-    fn search_args_accept_multiple_queries() {
-        let args: SearchArgs = serde_json::from_value(json!({
-            "queries": ["alpha", "needle"],
-            "case_sensitive": false
-        }))
-        .expect("multi-query args should parse");
-
-        let request = args.into_request();
-
-        assert_eq!(
-            request,
-            SearchMemoriesRequest {
-                queries: vec!["alpha".to_string(), "needle".to_string()],
-                match_mode: SearchMatchMode::Any,
-                path: None,
-                cursor: None,
-                context_lines: 0,
-                case_sensitive: false,
-                normalized: false,
-                max_results: DEFAULT_SEARCH_MAX_RESULTS,
-            }
-        );
-    }
-
-    #[test]
-    fn search_args_accept_windowed_all_match_mode() {
-        let args: SearchArgs = serde_json::from_value(json!({
-            "queries": ["alpha", "needle"],
-            "match_mode": {
-                "type": "all_within_lines",
-                "line_count": 3
-            }
-        }))
-        .expect("windowed all args should parse");
-
-        let request = args.into_request();
-
-        assert_eq!(
-            request,
-            SearchMemoriesRequest {
-                queries: vec!["alpha".to_string(), "needle".to_string()],
-                match_mode: SearchMatchMode::AllWithinLines { line_count: 3 },
-                path: None,
-                cursor: None,
-                context_lines: 0,
-                case_sensitive: true,
-                normalized: false,
-                max_results: DEFAULT_SEARCH_MAX_RESULTS,
-            }
-        );
-    }
-
-    #[test]
-    fn search_args_reject_legacy_single_query() {
-        let err = serde_json::from_value::<SearchArgs>(json!({
-            "query": "needle",
-        }))
-        .expect_err("legacy query field should be rejected");
-
-        assert!(err.to_string().contains("unknown field"));
-        assert!(err.to_string().contains("query"));
     }
 }
